more pushed code

+++ 08_SI_Calculator.py
@@ -1,8 +1,8 @@
 # Here we are creating SI Calculator using input function and variable.
 
-p = int(input("Enter your principle amount:- "))
-r = int(input("Enter your rate of interest:- "))
-t = int(input("Enter your time duration:- "))
+p = float(input("Enter your principle amount:- "))
+r = float(input("Enter your rate of interest:- "))
+t = float(input("Enter your time duration:- "))
 
 si = (p*r*t)/100
